requirements.txt
easyocr==1.7.2
inference==0.27.0
numpy==2.1.3
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
python-dotenv==1.0.1
supervision==0.25.0
